Add files via upload
Printed standing up, they are a snug fit so you won't be able to push them in by hand, at least I was not able to. I used a rubber hammer and a small piece of wood on the part to push it in.

I use M3x16mm screws to secure them in with the skirts, as shown in the picture.